Added enemy's health bar. Added to the player's ability haduken remove health in enemy.

+++ Assets/_scripts/Player/PlayerHealth.cs
@@ -5,7 +5,7 @@ using System.Collections;
 public class PlayerHealth : MonoBehaviour
 {
 	public int startingHealth = 100;                            // The amount of health the player starts the game with.
-	public int currentHealth;                                   // The current health the player has.
+	public float currentHealth;                                   // The current health the player has.
 	public Slider healthSlider;                                 // Reference to the UI's health bar.
 //	public Image damageImage;                                   // Reference to an image to flash on the screen on being hurt.
 //	public AudioClip deathClip;                                 // The audio clip to play when the player dies.
@@ -54,7 +54,7 @@ public class PlayerHealth : MonoBehaviour
 	}
 	
 	
-	public void TakeDamage (int amount)
+	public void TakeDamage (float amount)
 	{
 		// Set the damaged flag so the screen will flash.
 //		damaged = true;
